Clean data, create figures
I cleaned the data with the added hair color and BMI variables, removing entries with abnormal and NA values in one or more variables. Then, I ran code creating a series of graphs based off the data, including original code to make a boxplot showing height distributions by hair color and a scatterplot showing BMI as a function of weight.